Fixed link app-infra-config
