+++ app/src/main/kotlin/fr/sacane/jmanager/app/WebConfig.kt
@@ -1,0 +1,13 @@
+package fr.sacane.jmanager.app
+
+import fr.sacane.jmanager.server.adapters.LoginTransactionAdapter
+import fr.sacane.jmanager.server.repositories.UserRepository
+import fr.sacane.jmanager.server.spring.SpringLayerService
+import org.springframework.context.annotation.ComponentScan
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.PropertySource
+
+@Configuration
+@PropertySource("classpath:application.properties")
+class WebConfig {
+}
