+++ infra/src/main/kotlin/fr/sacane/jmanager/InfraConfig.kt
@@ -1,0 +1,9 @@
+package fr.sacane.jmanager
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableAutoConfiguration
+class InfraConfig {
+}
