--- infra/src/main/kotlin/fr/sacane/jmanager/app/README.md
@@ -1,4 +1,0 @@
-# User-side
-
----
-_Details coming soon..._
